+++ bin/update-get-release-tag.mjs
@@ -1,0 +1,28 @@
+#!/usr/bin/env node
+let shutdown_handler = null;
+try {
+    shutdown_handler = await (await import("../../flux-shutdown-handler-api/src/Adapter/Api/ShutdownHandlerApi.mjs")).ShutdownHandlerApi.new()
+        .getShutdownHandler();
+
+    const path = process.argv[2] ?? null;
+    if (path === null) {
+        throw new Error("Please pass a path");
+    }
+
+    process.stdout.write(await (await import("../src/Adapter/Api/PublishUtilsApi.mjs")).PublishUtilsApi.new(
+        shutdown_handler
+    )
+        .updateGetReleaseTag(
+            path
+        ));
+} catch (error) {
+    console.error(error);
+
+    if (shutdown_handler !== null) {
+        await shutdown_handler.shutdown(
+            1
+        );
+    } else {
+        process.exit(1);
+    }
+}